==> notebooks/02_clean_feature.ipynb <==
# Data Cleaning & Feature Engineering

The goal of this notebook is to clean and transform the data into the panel format ready for analysis

import pandas as pd 
import numpy as np 
import seaborn as sns
import matplotlib.pyplot as plt 

df = pd.read_csv('../data/trips.csv')
df.head()

## 1. Data Cleaning

* Datatypes
* Missing Data
* Duplicates
* Outliers

# Summary statistics of the dataset
df.describe()

# Check datatypes
df.dtypes
df['start_time'] = pd.to_datetime(df['start_time'])

# Check for missing values
df.isna().sum() 

# Check for duplicates
df.duplicated().sum()

## 2. Feature Engineering

Some features already generated during the inital data pull.

Adding a few more features for further analysis

# Create holiday dummy variable based on US federal holidays 1 if not 0 
from pandas.tseries.holiday import USFederalHolidayCalendar
cal = USFederalHolidayCalendar()
holidays = cal.holidays(start='2018-01-01', end='2018-12-31')
df['start_time'] = pd.to_datetime(df['start_time'])
df['holiday'] = df['start_time'].dt.normalize().isin(holidays).astype(int)

# Create day of week variable as numeric and string
df['day_of_week_num'] = df['start_time'].dt.dayofweek
df['day_of_week'] = df['start_time'].dt.day_name()  

# Encode neighborhood names as numeric ids
neighborhoods = df['neighborhood_start'].unique()
neighborhood_to_id = {name: idx for idx, name in enumerate(neighborhoods)}
df['neighborhood_id'] = df['neighborhood_start'].map(neighborhood_to_id)

# Create a day variable include 365 days from Jan 1 to Dec 31
df['day'] = df['start_time'].dt.dayofyear   


# Set up panel data format with neighborhood and day of week as index
df_panel = df.groupby(['neighborhood_id', 'day']).agg({
    'trip_count': 'sum',
    'avg_trip_minutes': 'mean',
    'day_mean_temperature': 'mean',
    'day_mean_wind_speed': 'mean',
    'day_total_precipitation': 'mean',
    'holiday': 'max'  # if any day in the group is a holiday, mark the whole group as holiday
}).reset_index()

df_panel.to_csv('../data/trips_panel.csv', index=False)
df.to_csv('../data/trips_cleaned.csv', index=False)
df_panel.head().sort_values(by=['neighborhood_id', 'day'])

# display count of days per neighborhood    
df_panel.groupby('neighborhood_id')['day'].count()
